refactor(agent): remove legacy auto-memory integration
- remove legacy auto-memory prompt injection and workspace memory file context
- move AGENTS/CLAUDE instruction loading into a standalone instruction_context service
- drop workspace_memory_files from runtime contracts, custom agent serialization, and UI config
- remove obsolete memory storage path exposure from runtime layout and desktop storage API
