+++ src/crates/assembly/core/src/agentic/agents/definitions/modes/claw.rs
@@ -74,7 +74,6 @@ impl Agent for ClawMode {
         UserContextPolicy::empty()
             .with_workspace_context()
             .with_workspace_instructions()
-            .with_workspace_memory_files()
     }
 
     fn is_readonly(&self) -> bool {
