+++ src/crates/assembly/core/src/agentic/agents/registry/custom.rs
@@ -482,9 +482,6 @@ impl AgentRegistry {
                     bitfun_agent_runtime::prompt::UserContextSection::WorkspaceInstructions => {
                         "workspace_instructions"
                     }
-                    bitfun_agent_runtime::prompt::UserContextSection::WorkspaceMemoryFiles => {
-                        "workspace_memory_files"
-                    }
                     bitfun_agent_runtime::prompt::UserContextSection::ProjectLayout => {
                         "project_layout"
                     }
