+++ src/crates/assembly/core/src/agentic/session/session_manager.rs
@@ -6063,7 +6063,7 @@ mod tests {
             .expect("session should be created");
         let identity = SystemPromptCacheIdentity::new("template:agentic_mode");
         let user_context_identity = UserContextCacheIdentity::new(
-            "workspace_context|workspace_instructions|workspace_memory_files|project_layout",
+            "workspace_context|workspace_instructions|project_layout",
         );
 
         manager
@@ -6278,7 +6278,7 @@ mod tests {
             .expect("session should be created");
         let identity = SystemPromptCacheIdentity::new("template:agentic_mode");
         let user_context_identity = UserContextCacheIdentity::new(
-            "workspace_context|workspace_instructions|workspace_memory_files|project_layout",
+            "workspace_context|workspace_instructions|project_layout",
         );
 
         manager
@@ -6358,7 +6358,7 @@ mod tests {
             .expect("target session should be created");
         let identity = SystemPromptCacheIdentity::new("template:agentic_mode");
         let user_context_identity = UserContextCacheIdentity::new(
-            "workspace_context|workspace_instructions|workspace_memory_files|project_layout",
+            "workspace_context|workspace_instructions|project_layout",
         );
 
         manager
@@ -6439,7 +6439,7 @@ mod tests {
             .expect("session should be created");
         let identity = SystemPromptCacheIdentity::new("template:agentic_mode");
         let user_context_identity = UserContextCacheIdentity::new(
-            "workspace_context|workspace_instructions|workspace_memory_files|project_layout",
+            "workspace_context|workspace_instructions|project_layout",
         );
 
         manager
