+++ src/crates/assembly/core/src/infrastructure/app_paths/path_manager.rs
@@ -386,11 +386,6 @@ impl PathManager {
         self.project_runtime_root(workspace_path).join("plans")
     }
 
-    /// Get project memory directory: ~/.bitfun/projects/<workspace-slug>/memory/
-    pub fn project_memory_dir(&self, workspace_path: &Path) -> PathBuf {
-        self.project_runtime_root(workspace_path).join("memory")
-    }
-
     fn project_runtime_slug(&self, workspace_path: &Path) -> String {
         let requested_path = workspace_path.to_path_buf();
         if let Some(slug) = self.cached_project_runtime_slug(&requested_path) {
