--- src/crates/assembly/core/src/service/agent_memory/mod.rs
@@ -1,6 +1,0 @@
-mod auto_memory;
-mod instruction_context;
-
-pub(crate) use auto_memory::build_workspace_agent_memory_prompt;
-pub(crate) use auto_memory::build_workspace_memory_files_context;
-pub(crate) use instruction_context::build_workspace_instruction_files_context;
