+++ src/crates/assembly/core/src/service/instruction_context.rs
