+++ src/crates/assembly/core/src/service/mod.rs
@@ -4,8 +4,6 @@
 //! isolated. High-coupling runtime services stay here until their port
 //! contracts and equivalence tests are explicit.
 
-#[cfg(feature = "product-full")]
-pub(crate) mod agent_memory; // Agent memory prompt helpers
 #[cfg(feature = "service-integrations")]
 pub mod announcement; // Announcement / feature-demo / tips system
 pub(crate) mod bootstrap; // Workspace persona bootstrap helpers
@@ -16,6 +14,8 @@ pub mod filesystem; // FileSystem management
 #[cfg(feature = "service-integrations")]
 pub mod git; // Git service
 pub mod i18n; // I18n service
+#[cfg(feature = "product-full")]
+pub(crate) mod instruction_context; // Workspace instruction file prompt helpers
 pub mod lsp; // LSP (Language Server Protocol) system
 #[cfg(feature = "service-integrations")]
 pub mod mcp; // MCP (Model Context Protocol) system
