+++ src/crates/execution/agent-runtime/src/prompt.rs
@@ -450,7 +450,6 @@ fn computer_use_key_chord_guidance(host_os: &str) -> &'static str {
 pub enum UserContextSection {
     WorkspaceContext,
     WorkspaceInstructions,
-    WorkspaceMemoryFiles,
     ProjectLayout,
 }
 
@@ -486,10 +485,6 @@ impl UserContextPolicy {
         self.with_section(UserContextSection::WorkspaceInstructions)
     }
 
-    pub fn with_workspace_memory_files(self) -> Self {
-        self.with_section(UserContextSection::WorkspaceMemoryFiles)
-    }
-
     pub fn with_project_layout(self) -> Self {
         self.with_section(UserContextSection::ProjectLayout)
     }
@@ -522,7 +517,6 @@ impl UserContextSection {
         match self {
             Self::WorkspaceContext => "workspace_context",
             Self::WorkspaceInstructions => "workspace_instructions",
-            Self::WorkspaceMemoryFiles => "workspace_memory_files",
             Self::ProjectLayout => "project_layout",
         }
     }
